refactor: Move constants into src/constants
no code change
