+++ src/components/EditDialogs/PurposeEditDialog/PurposeEditDialog.jsx
@@ -8,7 +8,7 @@ import { purposes } from 'src/components/helpers'
 import { createGeojsonWithModifiedPurpose } from 'src/components/EditDialogs/PurposeEditDialog/helpers'
 import { PurposeAvatar } from 'src/components/Avatar'
 import { useTrip } from 'src/components/Providers/TripProvider'
-import { OTHER_PURPOSE } from 'src/constants/const'
+import { OTHER_PURPOSE } from 'src/constants'
 
 const makeOptions = t => {
   const options = purposes.map(purpose => ({
